Move ModelData to non-client package
