+++ patches/net/minecraft/client/renderer/block/model/BakedQuad.java.patch
@@ -7,7 +7,7 @@
 +/**
 + * @param hasAmbientOcclusion {@code false} to force-disable AO for this quad
 + *                            or {@code true} to use the behavior dictated by
-+ *                            {@link net.neoforged.neoforge.client.extensions.IBakedModelExtension#useAmbientOcclusion(net.minecraft.world.level.block.state.BlockState, net.neoforged.neoforge.client.model.data.ModelData, net.minecraft.client.renderer.RenderType)}
++ *                            {@link net.neoforged.neoforge.client.extensions.IBakedModelExtension#useAmbientOcclusion(net.minecraft.world.level.block.state.BlockState, net.neoforged.neoforge.model.data.ModelData, net.minecraft.client.renderer.RenderType)}
 + */
  @OnlyIn(Dist.CLIENT)
 -public record BakedQuad(int[] vertices, int tintIndex, Direction direction, TextureAtlasSprite sprite, boolean shade, int lightEmission) {
