+++ patches/net/minecraft/client/renderer/chunk/RenderChunkRegion.java.patch
@@ -4,13 +4,13 @@
      private final int minChunkZ;
      protected final RenderChunk[] chunks;
      protected final Level level;
-+    private final it.unimi.dsi.fastutil.longs.Long2ObjectFunction<net.neoforged.neoforge.client.model.data.ModelData> modelDataSnapshot;
++    private final it.unimi.dsi.fastutil.longs.Long2ObjectFunction<net.neoforged.neoforge.model.data.ModelData> modelDataSnapshot;
  
 +    @Deprecated
      RenderChunkRegion(Level p_200456_, int p_200457_, int p_200458_, RenderChunk[] p_350680_) {
-+        this(p_200456_, p_200457_, p_200458_, p_350680_, net.neoforged.neoforge.client.model.data.ModelDataManager.EMPTY_SNAPSHOT);
++        this(p_200456_, p_200457_, p_200458_, p_350680_, net.neoforged.neoforge.model.data.ModelDataManager.EMPTY_SNAPSHOT);
 +    }
-+    RenderChunkRegion(Level p_200456_, int p_200457_, int p_200458_, RenderChunk[] p_350680_, it.unimi.dsi.fastutil.longs.Long2ObjectFunction<net.neoforged.neoforge.client.model.data.ModelData> modelDataSnapshot) {
++    RenderChunkRegion(Level p_200456_, int p_200457_, int p_200458_, RenderChunk[] p_350680_, it.unimi.dsi.fastutil.longs.Long2ObjectFunction<net.neoforged.neoforge.model.data.ModelData> modelDataSnapshot) {
          this.level = p_200456_;
          this.minChunkX = p_200457_;
          this.minChunkZ = p_200458_;
@@ -31,7 +31,7 @@
 +    }
 +
 +    @Override
-+    public net.neoforged.neoforge.client.model.data.ModelData getModelData(BlockPos pos) {
++    public net.neoforged.neoforge.model.data.ModelData getModelData(BlockPos pos) {
 +        return modelDataSnapshot.get(pos.asLong());
 +    }
 +
