+++ src/client/java/net/neoforged/neoforge/client/RenderTypeHelper.java
@@ -12,7 +12,7 @@ import net.minecraft.client.renderer.Sheets;
 import net.minecraft.client.renderer.block.model.BlockStateModel;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.block.state.BlockState;
-import net.neoforged.neoforge.client.model.data.ModelData;
+import net.neoforged.neoforge.model.data.ModelData;
 
 /**
  * Provides helper functions replacing those in {@link ItemBlockRenderTypes}.
